+++ src/main/java/tekcays_addon/api/metatileentity/predicates/Predicates.java
@@ -1,4 +1,4 @@
-package tekcays_addon.api.metatileentity;
+package tekcays_addon.api.metatileentity.predicates;
 
 import gregtech.api.pattern.TraceabilityPredicate;
 
@@ -11,6 +11,16 @@ public class Predicates {
         return new TraceabilityPredicate((blockWorldState) -> {
             if (air().test(blockWorldState)) {
                 blockWorldState.getMatchContext().increment(matchContext, 1);
+                return true;
+            } else
+                return false;
+        });
+    }
+
+    public static TraceabilityPredicate heightIndicatorPredicate(TraceabilityPredicate predicate, String matchContext, int incrementValue) {
+        return new TraceabilityPredicate((blockWorldState) -> {
+            if (predicate.test(blockWorldState)) {
+                blockWorldState.getMatchContext().increment(matchContext, incrementValue);
                 return true;
             } else
                 return false;
