+++ src/main/java/tekcays_addon/common/metatileentities/multi/MetaTileEntityElectricConverter.java
@@ -33,6 +33,8 @@ import javax.annotation.Nonnull;
 import javax.annotation.Nullable;
 import java.util.List;
 
+import static tekcays_addon.gtapi.consts.TKCYAValues.ROOM_TEMPERATURE;
+
 public class MetaTileEntityElectricConverter extends ModulableRecipeMapController implements IHeatingCoil {
 
     private int blastFurnaceTemperature;
@@ -87,6 +89,7 @@ public class MetaTileEntityElectricConverter extends ModulableRecipeMapControlle
 
      */
 
+    @Nonnull
     @Override
     protected BlockPattern createStructurePattern() {
         return FactoryBlockPattern.start()
