+++ src/main/java/tekcays_addon/common/metatileentities/multi/storage/MetaTileEntityModulableMultiblockTank.java
@@ -36,7 +36,7 @@ import java.util.List;
 
 import static gregtech.api.unification.material.Materials.TreatedWood;
 import static gregtech.api.util.RelativeDirection.*;
-import static tekcays_addon.api.metatileentity.Predicates.isAir;
+import static tekcays_addon.api.metatileentity.predicates.Predicates.isAir;
 import static tekcays_addon.api.metatileentity.TankMethods.*;
 
 public class MetaTileEntityModulableMultiblockTank extends MultiblockWithDisplayBase {
