--- src/main/java/tekcays_addon/gtapi/metatileentity/multiblock/HeatedPressureContainerMultiblockController.java
@@ -1,54 +1,0 @@
-package tekcays_addon.gtapi.metatileentity.multiblock;
-
-
-import gregtech.api.capability.impl.MultiblockRecipeLogic;
-import gregtech.api.metatileentity.multiblock.RecipeMapMultiblockController;
-import gregtech.api.pattern.TraceabilityPredicate;
-import gregtech.api.recipes.RecipeMap;
-import lombok.Getter;
-import net.minecraft.util.ResourceLocation;
-import tekcays_addon.api.recipe.PressureContainerCheckRecipeHelper;
-import tekcays_addon.gtapi.capability.containers.IContainerDetector;
-import tekcays_addon.gtapi.capability.containers.IHeatContainer;
-import tekcays_addon.gtapi.capability.containers.IPressureContainer;
-import tekcays_addon.gtapi.capability.list.HeatContainerList;
-import tekcays_addon.gtapi.capability.machines.IContainerDetectorMachine;
-import tekcays_addon.gtapi.capability.machines.IHeatMachine;
-import tekcays_addon.gtapi.capability.machines.IPressureMachine;
-
-import java.util.List;
-
-import static tekcays_addon.gtapi.metatileentity.multiblock.TKCYAMultiblockAbility.*;
-
-public abstract class HeatedPressureContainerMultiblockController extends RecipeMapMultiblockController implements PressureContainerCheckRecipeHelper, IPressureMachine, IHeatMachine, IContainerDetectorMachine {
-
-    @Getter
-    protected IPressureContainer pressureContainer;
-    @Getter
-    protected IHeatContainer heatContainer;
-    @Getter
-    protected IContainerDetector containerDetector;
-    @Getter
-
-    protected int volume;
-
-
-    public HeatedPressureContainerMultiblockController(ResourceLocation metaTileEntityId, RecipeMap<?> recipeMap) {
-        super(metaTileEntityId, recipeMap);
-        this.recipeMapWorkable = new MultiblockRecipeLogic(this);
-    }
-
-    @Override
-    protected void initializeAbilities() {
-        super.initializeAbilities();
-        this.heatContainer = new HeatContainerList(getAbilities(HEAT_CONTAINER));
-
-        List<IPressureContainer> list = getAbilities(PRESSURE_CONTAINER);
-        this.pressureContainer = list.isEmpty() ? null : list.get(0);
-
-        List<IContainerDetector> list2 = getAbilities(CONTAINER_CONTROL);
-        this.containerDetector = list2.isEmpty() ? null : list2.get(0);
-    }
-
-
-}
